# fix(relationships): remove legacy document storage routing from relationship queries